+++ code/data-service/src/main/java/ldn/cs/fusion/dao/StaffDao.java
@@ -27,9 +27,9 @@ public interface StaffDao {
     int getTotalStaff(@Param("statement") String statement, @Param("types") int types);
 
     //granularity: 1、年 2、季度 3、月
-    List<Person> getPersonInfos(@Param("time") int time, @Param("granularity") int granularity);
+    List<Person> getPersonInfos(@Param("time") long time, @Param("granularity") int granularity);
 
-    List<Position> getPositionInfos(@Param("time") int time, @Param("granularity") int granularity);
+    List<Position> getPositionInfos(@Param("time") long time, @Param("granularity") int granularity);
 
-    List<Skill> getSkillInfos(@Param("time") int time, @Param("granularity") int granularity);
+    List<Skill> getSkillInfos(@Param("time") long time, @Param("granularity") int granularity);
 }
